[ListSort, Rename] Rename ListSortHelper to PackageTableSortHelper
Rename to reflect broader scope — the helper sorts package table rows
for any table-based command (list, search, upgrade), not just list.
