+++ src/AppInstallerCLICore/Workflows/WorkflowBase.cpp
@@ -3,7 +3,7 @@
 #include "pch.h"
 #include "WorkflowBase.h"
 #include "ExecutionContext.h"
-#include "ListSortHelper.h"
+#include "PackageTableSortHelper.h"
 #include "PromptFlow.h"
 #include "ShowFlow.h"
 #include "Sixel.h"
